Move tests to repo
